RP-107 Change over some of the functions to JSON.js and add event handlers
